Safari-611.1.10/LayoutTests/inspector/controller/runtime-controller-import-expected.txt

CONSOLE MESSAGE: TypeError: Module specifier, '' does not start with "/", "./", or "../". Referenced from runtime-controller-import.html
CONSOLE MESSAGE: TypeError: Importing a module script failed.
CONSOLE MESSAGE: Cocoa is Sweet.
CONSOLE MESSAGE: %o
CONSOLE MESSAGE: Cocoa is Sweet. Cappuccino is Awesome.
CONSOLE MESSAGE: %o
Tests for RuntimeManager operations.


== Running test suite: RuntimeManager
-- Running test case: RuntimeManager.prototype.evaluateInInspectedWindow.ImportFromConsole
Source: await import('')
PASS: Transformed. Should log the value or an exception.
Source: await import('./not-found.js')
PASS: Transformed. Should log the value or an exception.
Source: await import('./resources/cocoa.js').then((cocoa) => { console.log(`${cocoa.name} is ${cocoa.taste}.`); })
PASS: Transformed. Should log the value or an exception.
Source: await import('./resources/drink.js').then((ns) => { console.log(`${ns.Cocoa.name} is ${ns.Cocoa.taste}. ${ns.Cappuccino.name} is ${ns.Cappuccino.taste}.`); })
PASS: Transformed. Should log the value or an exception.
